aml_squad_app/requirements.txt
streamlit==1.17.0
fastapi==0.89.1
uvicorn==0.20.0
python-multipart==0.0.5
tabula-py==2.6.0
azure-ai-formrecognizer==3.2.0
PyPDF2==3.0.1